Basic Struct Generation
This commit adds tooling for generating structs (records in gir
terminology). For 'RefStruct' types, where the struct is a simple
one to one marshal from C, it also generates the struct fields
automatically.

Finally, it continues to create classes for structs that cannot be
described by intropsection data (e.g. they are disguised, etc). In
future, this should be extended to use IntPtr automatically.

Next Steps:
 - Improve 'resolve type' to work for fields as well as methods