Generate Signals Properly
This commit allows for signals to be generated with the correct
syntax, and with the entire project compiling again. The codegen
is itself complete (minus some edge-case features that we may want
to consider down the road).

The final remaining step is to generate custom SignalArgs-based types
for signals with additional parameters.