SignalArgs Parameter Generation
This takes SignalArgs generation a step further, automatically
generating a field for each property. The final step is to be able
to marshal between GValue and the given field's concrete type.

We might want to consider avoiding GValue and Closures altogether if
we go down this route, since it adds an additional layer of overhead
with seemingly few benefits.