Implement Fuzzy Search
This commit implements a string-based
lookup for GTypes based on matching
loaded assemblies with string
representations of GTypes.

It is much more "correct" than our
previous approach, but still has several
outstanding issues which would be better
served by moving away from the runtime
type dictionary paradigm.

This should be considered an interim
approach to resolving some major type
dictionary issues, and not in any way
permanent.

We should aim to replace this with a
compile-time generated typemap akin to
Xamarin for both performance and
correctness.