GStreamer + Plugins
This commit adds more classes to GStreamer, as well
as starts generating the following plugins:
 * Gst.Audio
 * Gst.Base
 * Gst.Pbutils
 * Gst.Video

There is a lot of temporary glue present in GObject
to facilitate this. Before merging, we will need to
decide on a more permanent API for both GObject and
GStreamer.

Additionally, due to issues with the generator,
there is a lot of boilerplate here that should either
be auto generated, or avoided entirely (e.g. using
ref structs).