TypeDict: Cache fallback type queries
In addition to recursively registering types when a lookup fails, we also
want to cache the failed query so that future lookups only need a simple
dictionary lookup, rather than the more expensive fuzzy assembly search
procedure.

This also includes a bugfix to logging, where the incorrect gtype was
being printed to console upon fallback.

Fixes #158