Create a Service Manager
Refactors the test service created in the previous commit
to be part of a service manager, which allows services to
be registered beforehand and obtained at runtime.

Services are small utility classes which provide singular
functions. Multiple services should be combined to create
more advanced functionality. The idea behind this is to
keep all logic for a particular feature in one place, as
to prevent accidental breakage.