Split Generator into Generator and Repository
Let's make a data driven generator. This commit takes
out the finished-half of the generator and splits it
into a separate Repository project.

This includes a high level Repository construct to
import GIR files, and a clear separation between
Serialised XML and the clean, processed data.

The idea is that the Repository will handle references,
types, etc, so the generator can focus on doing codegen.

Due to the stable nature of GLib/GObject, this allows us
to effectively decouple Generation from file format.