Move TypeDict to Repository + Various Fixes, Cleanup
We might be able to get away with resolving our
references ahead of time, allowing the generator
to have a much cleaner, non-implementation dependent
structure.

This involves moving the type dict to Repository,
populating it, and iterating over references using
TypeDictView to resolve strings.

This commit only moves TypeDict and TypeDictView over,
no actual reference resolution happens.

It also contains other fixes and changes to get
the Loader and Parser to work. XmlSerializer requires
that classes are public to work, so revert the change
that made them internal.