String-based Type Resolution
Gir XML Data uses strings to reference other types both
within and external to the namespace. This commit uses
a type dictionary and corresponding view into the dictionary
to accurately resolve and reference other symbols.

It expands the parser and adds a minimal set of records
to satisfy the symbol resolver. Pending some additions to
available public data, this should be largely sufficient
for the proof-of-concept generator.