Remove ServiceManager
Service Managers are considered an anti-pattern. As
we know all services at compile time and no longer
have state to manage, there is no longer a reason to
use one.

This simplifies the writer by removing the service
manager and simply accessing them as private fields.

We may want to make services static in the future (i.e.
pure functions).