Replace TypeReference with it's interface.
This removes the codes "expectation" to use a TypeReference class. Now it just expects an interface of ITypeReference.

This moves the control of the instantiated class for this interface into a single other class: The TypeReferenceFactory.

The code is now independent of the TypeReference class but only dependent on it's interface.