Refactored model: - Refactored implementations of IType into classes as those describe a certain type and should only exist once (no value based semantics). - Refactored types which do not describe distinct "objects" into records: Method, ReturnValue
A "BasicType" could have a NativeName but no ManagedName. Another instance could have both properties set in our model tree. Both instances should be reflected by same instance to avoid ambiguity. In this way we can be sure that changing a type changes all other places in the model tree which use this type.

Records with it's value based semantics can hide errors or lead to false assumptions if used for types which are referenced in several places in the code as we expect them to always contain identical data.