MethodFactory: Support moved methods
The factory requires a MethodInfo to have either a name or the moved to field set. As the model does not support "moved" methods those are handled via a custom Exception.

If a method has no name and did not move this is a general exception which hints to an error in the gir file.