GObject: Implement memory management
- Fixes GdkPixbuf compilation errors
- Adds GdkPixbuf sample to solution
- Implement ToggleRef for all objects using ObjectHandle

The ToggleRef instance is taking control of the reference given to it. To release a reference which was passed to a toggle ref the toggle ref must be disposed.

The ObjectHandle is automatically mapping the given instance and handle. The ObjectMapper itself makes sure that the mapping happens via a ToggleRef which controls the object reference.

If the ObjectHandle is unmapping itself the ObjectMapper makes sure to dispose the ToggleRef and thus freeing unmanaged data in the C world.