GirLoader: Start moving TypeInformation data into the CType itself
The CType itself carries this information in the input model, too.

If the data is not located in the CType itself it is not possible to have a Type specifically for e.g. a "guint8*" which represents a string. The original code would create a "CTypeName" of "guint" (stripping the "*") which results in a resolved type of "sbyte" instead of "string".