TypeRegistration: Disable default class and instance init functions
If multiple instances of the same type are created, the instances of the delegates may be already garbage collected and the runtime will crash. Therefor the default handler are not set anymore to avoid a callback into a garbage collected instance of a delegate.

This is only temporary until the typeInfo object is properly cached to avoid losing the ClassInit and InstanceInit delegate instances.