Removed fallback return type renderer
The fallback return type renderer has been removed, so that method
rendering in interfaces is more consistent with the rendering process in
classes. Previously, the fallback return type renderer would allow some
methods to be rendered in an interface but not in implementations of
that interface, which resulted in broken bindings.