Support callbacks returning a string with transfer none This introduces a new public class "ConstantString". This class can be used to keep a pointer to a string alive as long as the instance is alive.
In case of callbacks a returned string with transfer none should live forever. A user can create an instance of "ConstantString" and keep the instance around forever to keep the pointer alive forever.

This puts the burden of managing string memory which lives forever on the user side.

If the string would simply allocated by gir.core and never be freed there could be huge memory leaks if strings are allocated again and again and they would never be freed.