Prefer zero terminated string array over length based arrays.
As zero terminated string arrays require a certain memory layout (NULL at the end) there could be a problem if a method has an array parameter which is zero terminated and length based at the same time. Preferring zero terminated rendering ensures that the memory layout fulfills the requirements.

Additionally this fixes a bug that the zero terminated flag was never actively set as bool.TryParse can't parse "0" or "1" but the serializer does.