Object: Remove INotifyPropertyChanged interface
Implementing INotifyPropertyChanged seems not to be helpful currently. The interface was implemented since the very beginning of GirCore but the PropertyChanged event was never raised. As there were no complaints there is probably no need for this interface.

Implementing it would mean that every object has 2 different ways of notifying about property changes (1. Notify event, 2. PropertyChanged) which seems a bit odd in the first place. As the notify event is the native system it should be preferred.

While considering a possible implementation it came to mind that the PropertyChanged event should contain the managed name of the property. This would be the only sensible way for a C# dev, so the instance could potentially be compatible to other frameworks which require an instance of INotifyPropertyChanged and access the property via reflection.

To raise the PropertyChanged event every object would always be required to subscribe to the Notify signal and forward it to PropertyChanged (after some property name mapping happend, which could get non trivial, too). If a user does need the INotifyPropertyChanged interface it could be implemented manually as a subclass and the mapping could be implemented manually.