Null-terminated array by convention
I think this partially resolves gircore/gir.core#955

At this moment, the code generated by the API for size-based arrays is
not functional.

Arrays used as return values are almost always NULL-terminated. This is
not mandatory, but it is the convention. So treating them like
NULL-terminated arrays will works like a charm 99% of the time.